feat: enhance ability to recover from crashes - Real time service won't be affected by restarting of other services - Pending song will be moved to playing if there is no playing song
