+++ server/src/real-time-radio/real-time-songs/real-time-songs.module.ts
@@ -1,14 +1,16 @@
-import { Module, OnModuleInit } from '@nestjs/common';
+import { forwardRef, Module, OnModuleInit } from '@nestjs/common';
 import { PrismaModule } from '../../prisma/prisma.module';
-import { RealTimeSongUtilsService } from './real-time-song-utils.service';
-import { RealTimeSongsService } from './real-time-songs.service';
+import { RealTimeStationsModule } from '../real-time-stations/real-time-stations.module';
+import { RealTimeSongsWorkerService } from './real-time-songs-worker.service';
+import { RealTimeSongService } from './real-time-songs.service';
 
 @Module({
-  imports: [PrismaModule],
-  providers: [RealTimeSongsService, RealTimeSongUtilsService],
+  imports: [PrismaModule, forwardRef(() => RealTimeStationsModule)],
+  providers: [RealTimeSongsWorkerService, RealTimeSongService],
+  exports: [RealTimeSongService],
 })
 export class RealTimeSongsModule implements OnModuleInit {
-  constructor(private readonly songsService: RealTimeSongsService) {}
+  constructor(private readonly songsService: RealTimeSongsWorkerService) {}
   async onModuleInit() {
     await this.songsService.scanAllPlayingSongsOnInitialization();
     this.songsService.subscribeSong();
