+++ server/src/real-time-radio/real-time-songs/real-time-songs.service.ts
@@ -1,95 +1,75 @@
-import { Injectable, Logger, InternalServerErrorException } from '@nestjs/common';
-import { Song } from '../../prisma/prisma.binding';
-import { RealTimeSongUtilsService } from './real-time-song-utils.service';
+import { Injectable, InternalServerErrorException, Logger } from '@nestjs/common';
+import { Song, SongSubscriptionPayload, SongSubscriptionWhereInput } from '../../prisma/prisma.binding';
+import { PrismaService } from '../../prisma/prisma.service';
 
 @Injectable()
-export class RealTimeSongsService {
-  /**
-   * @description A hash map store the current timeout instance of a station.
-   * This will be use to interrupt the current playing song of station
-   */
-  private playersMap: { [key: string]: NodeJS.Timeout } = {};
+export class RealTimeSongService {
+  private readonly logger = new Logger(RealTimeSongService.name);
+  constructor(private readonly prisma: PrismaService) {}
 
-  private readonly logger = new Logger(RealTimeSongsService.name);
-  constructor(private readonly utils: RealTimeSongUtilsService) {}
-
-  /**
-   * @description: Subscribe song event from Prisma to catch what event to do next
-   */
-  async subscribeSong() {
-    const subscription = await this.utils.getSongSubscription();
-    while (true) {
-      const { value } = await subscription.next();
-      const { song } = value;
-      switch (song.mutation) {
-        case 'CREATED':
-          await this.onSongCreated(song.node);
-          break;
-        case 'UPDATED':
-          await this.onSongUpdated(song.node);
-          break;
-        case 'DELETED':
-          await this.onSongDeleted(song.node);
-          break;
-      }
-    }
+  async getSongSubscription(args?: { where?: SongSubscriptionWhereInput }) {
+    const info = `{
+      mutation
+      node ${this.query}
+      previousValues { id createdAt updatedAt title url thumbnail duration status startedAt }
+    }`;
+    return this.prisma.subscription.song<{ song: SongSubscriptionPayload }>(args, info);
   }
 
   /**
-   * @description: Scan all songs which has **PLAYING** status to update the player status on module initialization
+   * TODO: Implement an algorithm to figure out which song will be played next
    */
-  async scanAllPlayingSongsOnInitialization() {
-    const songs = await this.utils.findAllPlayingSongs();
-    await Promise.all(songs.map(song => this.playSong(song)));
+  async findNextPlayingSongInStation(stationId: string): Promise<Song | null> {
+    const pendingSongs = await this.prisma.query.songs(
+      { where: { station: { id: stationId }, status: 'PENDING' } },
+      this.query,
+    );
+    if (pendingSongs.length > 0) {
+      return pendingSongs[0];
+    }
+    return null;
   }
 
-  protected async playSong(song: Song) {
-    if (!song.startedAt) {
-      throw new InternalServerErrorException(`Song does not include property startedAt`);
-    }
-
-    const startedAtDate = song.startedAt instanceof Date ? song.startedAt : new Date(song.startedAt);
-    const startedAt = startedAtDate.getTime();
-    const now = Date.now();
-    const duration = song.duration - (now - startedAt);
-    if (duration <= 0) {
-      await this.updatePlayedSongAndPlayNextSong(song);
-    } else {
-      this.logger.debug(`Song [${song.id}] will be timed out in ${duration} milliseconds`);
-      this.playersMap[song.station.id] = setTimeout(() => this.updatePlayedSongAndPlayNextSong(song), duration);
-    }
+  async findAllPlayingSongs() {
+    return this.prisma.query.songs({ where: { status: 'PLAYING' } }, this.query);
   }
 
-  protected async updatePlayedSongAndPlayNextSong(song: Song) {
-    await this.utils.updateSongStatusToPlayed(song.id);
-    delete this.playersMap[song.station.id];
-    let nextSong = await this.utils.findNextPlayingSongInStation(song.station.id);
-    if (nextSong) {
-      nextSong = await this.utils.updateSongStatusToPlaying(nextSong.id);
-      this.playSong(nextSong);
+  async findPlayingSongInStation(stationId: string): Promise<Song | null> {
+    const playingSongs = await this.prisma.query.songs(
+      { where: { station: { id: stationId }, status: 'PLAYING' } },
+      this.query,
+    );
+    if (playingSongs.length > 0) {
+      return playingSongs[0];
     }
+    return null;
   }
 
-  protected async onSongCreated(song: Song) {
-    this.logger.debug(`Song [${song.id}] "${song.title}" in station [${song.station.id}] has been created`);
-    const playingSong = await this.utils.findPlayingSongInStation(song.station.id);
-    if (!playingSong) {
-      song = await this.utils.updateSongStatusToPlaying(song.id);
-      await this.playSong(song);
+  async findSongOrFail(songId: string): Promise<Song> {
+    const song = await this.prisma.query.song({ where: { id: songId } }, this.query);
+    if (!song) {
+      throw new InternalServerErrorException(`Song ${songId} not found`);
     }
+    return song;
   }
 
-  protected async onSongUpdated(song: Song) {
-    this.logger.debug(
-      `Song [${song.id}] "${song.title}" in station [${song.station.id}] has been updated with status ${song.status}`,
+  async updateSongStatusToPlaying(songId: string) {
+    const startedAt = new Date();
+    return this.prisma.mutation.updateSong(
+      { where: { id: songId }, data: { status: 'PLAYING', startedAt } },
+      this.query,
     );
   }
 
-  protected async onSongDeleted(song: Song) {
-    this.logger.debug(`Song [${song.id}] "${song.title}" in station [${song.station.id}] has been deleted`);
-    if (song.status === 'PLAYING') {
-      clearTimeout(this.playersMap[song.station.id]);
-      delete this.playersMap[song.station.id];
-    }
+  async updateSongStatusToPlayed(songId: string) {
+    return this.prisma.mutation.updateSong({ where: { id: songId }, data: { status: 'PLAYED' } }, this.query);
   }
+
+  private query = `{
+    id createdAt updatedAt title url thumbnail duration status startedAt
+    creator { id }
+    station { id }
+    upVotes { id }
+    downVotes { id }
+  }`;
 }
