+++ server/src/real-time-radio/real-time-stations/real-time-stations.module.ts
@@ -1,7 +1,17 @@
-import { Module } from '@nestjs/common';
-import { RealTimeStationsService } from './real-time-stations.service';
+import { forwardRef, Module } from '@nestjs/common';
+import { PrismaModule } from '../../prisma/prisma.module';
+import { RealTimeSongsModule } from '../real-time-songs/real-time-songs.module';
+import { RealTimeStationsWorkerService } from './real-time-stations-worker.service';
+import { RealTimeStationService } from './real-time-stations.service';
 
 @Module({
-  providers: [RealTimeStationsService]
+  imports: [PrismaModule, forwardRef(() => RealTimeSongsModule)],
+  providers: [RealTimeStationsWorkerService, RealTimeStationService],
+  exports: [RealTimeStationService],
 })
-export class RealTimeStationsModule {}
+export class RealTimeStationsModule {
+  constructor(private readonly realTimeStationsService: RealTimeStationsWorkerService) {}
+  async onModuleInit() {
+    await this.realTimeStationsService.scanAllStationsOnInitialization();
+  }
+}
